fix: move selectors helper to separate module
