+++ lib/static/modules/load-plugin.js
@@ -80,7 +80,9 @@ async function initPlugin(plugin, pluginName) {
             const depArgs = deps.map(dep => whitelistedDeps[dep]);
             // cyclic dep, resolve it dynamically
             const actions = await import('./actions');
-            return plugin(...depArgs, {pluginName, actions});
+            const selectorsTree = await import('./selectors/tree');
+            const selectorsHelper = await import('./selectors/selectorsHelper');
+            return plugin(...depArgs, {pluginName, actions, selectors: {selectorsTree, selectorsHelper}});
         }
 
         return plugin;
