+++ lib/static/modules/selectors/tree.js
@@ -1,9 +1,11 @@
-import {last, initial, identity, flatMap, isEmpty, find, compact, findLast} from 'lodash';
+import {last, initial, identity, flatMap, compact, findLast} from 'lodash';
 import {createSelector} from 'reselect';
 import {getFilteredBrowsers, getTestNameFilter, getStrictMatchFilter, getViewMode} from './view';
 import viewModes from '../../../constants/view-modes';
 import {isFailStatus, isErroredStatus, isIdleStatus} from '../../../common-utils';
 import {isNodeFailed, isNodeSuccessful, isAcceptable} from '../utils';
+
+import {shouldSuiteBeShown, shouldBrowserBeShown} from './selectorsHelper';
 
 const getSuites = (state) => state.tree.suites.byId;
 const getBrowsers = (state) => state.tree.browsers.byId;
@@ -151,33 +153,9 @@ export const mkShouldSuiteBeShown = () => createSelector(
     shouldSuiteBeShown
 );
 
-function shouldSuiteBeShown(suite, suites, browsers, errorGroupBrowserIds, testNameFilter, strictMatchFilter, filteredBrowsers, viewMode) {
-    if (!isStatusMatchViewMode(suite.status, viewMode)) {
-        return false;
-    }
-
-    const suiteBrowsers = getSuiteBrowsers(suite, {suites, browsers});
-
-    return suiteBrowsers.some((browser) => {
-        const testName = browser.parentId;
-
-        if (!isTestNameMatchFilters(testName, testNameFilter, strictMatchFilter)) {
-            return false;
-        }
-
-        return shouldShowBrowser(browser, filteredBrowsers, errorGroupBrowserIds);
-    });
-}
-
 export const mkShouldBrowserBeShown = () => createSelector(
     getBrowserById, getResultStatus, getErrorGroupBrowserIds, getFilteredBrowsers, getViewMode,
-    (browser, lastResultStatus, errorGroupBrowserIds, filteredBrowsers, viewMode) => {
-        if (!isStatusMatchViewMode(lastResultStatus, viewMode)) {
-            return false;
-        }
-
-        return shouldShowBrowser(browser, filteredBrowsers, errorGroupBrowserIds);
-    }
+    shouldBrowserBeShown
 );
 
 function getSuiteResults(node, tree, filterFn = identity) {
@@ -196,7 +174,7 @@ function getSuiteResults(node, tree, filterFn = identity) {
     return flatMap(node.suiteIds, (suiteId) => getSuiteResults(suites[suiteId], tree, filterFn));
 }
 
-function getSuiteBrowsers(suite, tree) {
+export function getSuiteBrowsers(suite, tree) {
     const {suites, browsers} = tree;
 
     if (suite.browserIds) {
@@ -206,45 +184,17 @@ function getSuiteBrowsers(suite, tree) {
     return flatMap(suite.suiteIds, (suiteId) => getSuiteBrowsers(suites[suiteId], tree));
 }
 
-function isStatusMatchViewMode(status, viewMode) {
+export function isStatusMatchViewMode(status, viewMode) {
     if (viewMode === viewModes.ALL) {
         return true;
     }
 
+    // noinspection RedundantIfStatementJS
     if (viewMode === viewModes.FAILED && isFailStatus(status) || isErroredStatus(status)) {
         return true;
     }
 
     return false;
-}
-
-function isTestNameMatchFilters(testName, testNameFilter, strictMatchFilter) {
-    if (!testNameFilter) {
-        return true;
-    }
-
-    return strictMatchFilter ? testName === testNameFilter : testName.includes(testNameFilter);
-}
-
-export function shouldShowBrowser(browser, filteredBrowsers, errorGroupBrowserIds = []) {
-    if (isEmpty(filteredBrowsers) && isEmpty(errorGroupBrowserIds)) {
-        return true;
-    }
-
-    const browserToFilterBy = filteredBrowsers.length === 0 || find(filteredBrowsers, {id: browser.name});
-    const matchErrorGroup = errorGroupBrowserIds.length === 0 || errorGroupBrowserIds.includes(browser.id);
-
-    if (!browserToFilterBy || !matchErrorGroup) {
-        return false;
-    }
-
-    const browserVersionsToFilterBy = [].concat(browserToFilterBy.versions).filter(Boolean);
-
-    if (isEmpty(browserVersionsToFilterBy)) {
-        return true;
-    }
-
-    return browserVersionsToFilterBy.includes(browser.version);
 }
 
 export function getFailedSuiteResults(tree) {
